# evidencias/conformidade_2026-10-02/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de Agendamento de Consultas",
    "description": "Assessment: use somente dados ficticios neste ambiente local.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/consultas": {
      "post": {
        "tags": [
          "Consultas"
        ],
        "summary": "Criar Consulta",
        "operationId": "criar_consulta_consultas_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "JWTBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite configurado."
          },
          "401": {
            "description": "Credencial ausente, invalida ou revogada."
          },
          "403": {
            "description": "Papel ou vinculo nao autorizado."
          },
          "503": {
            "description": "Persistencia ou auditoria indisponivel."
          },
          "429": {
            "description": "Orcamento esgotado; consultar Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistencia ou estado."
          }
        }
      },
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Listar Consultas",
        "operationId": "listar_consultas_consultas_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsultaRead"
                  },
                  "title": "Response Listar Consultas Consultas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite configurado."
          },
          "401": {
            "description": "Credencial ausente, invalida ou revogada."
          },
          "403": {
            "description": "Papel ou vinculo nao autorizado."
          },
          "503": {
            "description": "Persistencia ou auditoria indisponivel."
          },
          "429": {
            "description": "Orcamento esgotado; consultar Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          }
        }
      }
    },
    "/consultas/{consulta_id}": {
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Obter Consulta",
        "operationId": "obter_consulta_consultas__consulta_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "exclusiveMinimum": 0,
              "title": "Consulta Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite configurado."
          },
          "401": {
            "description": "Credencial ausente, invalida ou revogada."
          },
          "403": {
            "description": "Papel ou vinculo nao autorizado."
          },
          "503": {
            "description": "Persistencia ou auditoria indisponivel."
          },
          "429": {
            "description": "Orcamento esgotado; consultar Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "404": {
            "description": "Consulta nao encontrada ou fora do ownership."
          }
        }
      },
      "put": {
        "tags": [
          "Consultas"
        ],
        "summary": "Atualizar Consulta",
        "operationId": "atualizar_consulta_consultas__consulta_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "exclusiveMinimum": 0,
              "title": "Consulta Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite configurado."
          },
          "401": {
            "description": "Credencial ausente, invalida ou revogada."
          },
          "403": {
            "description": "Papel ou vinculo nao autorizado."
          },
          "503": {
            "description": "Persistencia ou auditoria indisponivel."
          },
          "429": {
            "description": "Orcamento esgotado; consultar Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistencia ou estado."
          },
          "404": {
            "description": "Consulta nao encontrada ou fora do ownership."
          }
        }
      },
      "delete": {
        "tags": [
          "Consultas"
        ],
        "summary": "Excluir Consulta",
        "operationId": "excluir_consulta_consultas__consulta_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "exclusiveMinimum": 0,
              "title": "Consulta Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite configurado."
          },
          "401": {
            "description": "Credencial ausente, invalida ou revogada."
          },
          "403": {
            "description": "Papel ou vinculo nao autorizado."
          },
          "503": {
            "description": "Persistencia ou auditoria indisponivel."
          },
          "429": {
            "description": "Orcamento esgotado; consultar Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistencia ou estado."
          },
          "404": {
            "description": "Consulta nao encontrada ou fora do ownership."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsultaRead": {
        "properties": {
          "id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Id"
          },
          "paciente_id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Paciente Id"
          },
          "profissional_id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Profissional Id"
          },
          "data_hora": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora"
          },
          "status": {
            "$ref": "#/components/schemas/StatusConsulta"
          },
          "observacao": {
            "type": "string",
            "title": "Observacao"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "paciente_id",
          "profissional_id",
          "data_hora",
          "status",
          "observacao"
        ],
        "title": "ConsultaRead"
      },
      "ConsultaWrite": {
        "properties": {
          "paciente_id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Paciente Id"
          },
          "profissional_id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Profissional Id"
          },
          "data_hora": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora"
          },
          "status": {
            "$ref": "#/components/schemas/StatusConsulta",
            "default": "agendada"
          },
          "observacao": {
            "type": "string",
            "maxLength": 500,
            "title": "Observacao",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paciente_id",
          "profissional_id",
          "data_hora"
        ],
        "title": "ConsultaWrite",
        "description": "Contrato para criacao e substituicao completa (PUT)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StatusConsulta": {
        "type": "string",
        "enum": [
          "agendada",
          "cancelada",
          "realizada"
        ],
        "title": "StatusConsulta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      },
      "ClientBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "LaboratorioOAuth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "integracoes/token",
            "scopes": {
              "disponibilidade:read": "Ler somente horarios explicitamente ofertados."
            }
          }
        }
      },
      "JWTBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtido no login ou apos confirmar MFA."
      },
      "AgendaCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agenda_session",
        "description": "Cookie aceito somente no GET /agenda."
      }
    }
  }
}